fix circular dep
